Refactor product saving process to a use case
The save method in the ProductsController has been refactored to encapsulate the product saving logic into a new use case class, SaveProductUseCase. This new class contains the routine previously in the controller, and now the controller simply calls this use case. Additionally, unnecessary code was removed from ProductsCreate.vue, and a typo in Product.php was fixed.