Implement OneToOne relationship between User and UserProfile entities
This commit introduces a bi-directional OneToOne relationship between the User and UserProfile entities. It adds necessary methods to both User and UserProfile for handling this relationship. Additionally, a new database migration is included that modifies the userProfile table to add a foreign key to the User table. A minor change was made to the .idea/piniaClass.iml file removing a redundancy.