Add product category feature to the product entity
The commit introduces a new feature which allows categorization of products. A migration file has been created to setup the 'product_category' table and to add a 'product_category_id' foreign key to the 'product' table. New entity 'ProductCategory' has been created with related functionalities and repository. The 'Product' entity has been updated to include a reference to its category.