Add Symfony mailer and send welcome email on user registration, Define headers globally on useAPI composable.
This update integrates the Symfony mailer, enabling the system to send a welcome email to users upon registration. This is accomplished via changes in the 'SaveUserUseCase.php' where code for email sending was added. Corresponding dependencies were updated in 'composer.json' and 'composer.lock'. The changes also include a logout method for UserController for better handling of user authentication. Minor improvements to text in LoginUser.vue and CartStore.js are made to enhance clarity.