Add Symfony mailer and Twig template for welcome emails
The update includes the Symfony mailer integration to the backend, enabling the system to send welcome emails to users upon successful registration. The 'SaveUserUseCase.php' was updated to include this feature. Twig was also integrated to allow templated HTML emails, with corresponding updates made in 'composer.json', 'composer.lock' and the project settings. An example email template was also added.