Ofetch Interceptors -  Add error handling logic in API calls and redirect unauthenticated users
In this update, error handling logic has been added to the 'useApi' composable, logging the details of an error and redirecting users to the login page upon receiving a 401 status. Also, an unauthenticated check has been added to the global auth middleware, which redirects users to the login page if they are not authenticated. The 'Login' page has been assigned a new name, 'LoginUser'.