Add password hashing to AddUserCommand
In this update, the UserPasswordHasherInterface was added to the AddUserCommand's constructor. This allows us to hash the password when populating the 'user' entity, rather than inputting it in plaintext. As a result, this enhances the security of the application AND symfony wont let you log in if the password is not hashed.