Update Product entity and refactor serialization
The update primarily involves a refactor of how entities are serialized. Instead of using the JsonSerializable interface, the Symfony's Serializer component with groups attribute is now being utilized. This change enhances code organization, makes it easier to work with different views of the same resource, and ensures serialization responsibilities are handled by Symfony. A new enumeration was also created to manage the groupings.