Add PHPWord library and export functionality
The PHPWord library has been added to the project, and new export functionality has been provided in the "ExportDocumentsController". This allows documents to be exported in Word2007, ODF, and HTML formats. The document contents are hard-coded for now and the file is saved in the project root directory. The composer lock file has also been updated based on these additions.